Move APIs that are in Data package to API package
