+++ src/main/java/at/hannibal2/skyhanni/api/GuildAPI.kt
@@ -1,5 +1,6 @@
-package at.hannibal2.skyhanni.data
+package at.hannibal2.skyhanni.api
 
+import at.hannibal2.skyhanni.data.ProfileStorageData
 import at.hannibal2.skyhanni.events.LorenzChatEvent
 import at.hannibal2.skyhanni.utils.StringUtils.cleanPlayerName
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
