+++ src/main/java/at/hannibal2/skyhanni/api/PetAPI.kt
@@ -1,4 +1,6 @@
-package at.hannibal2.skyhanni.data
+package at.hannibal2.skyhanni.api
+
+import at.hannibal2.skyhanni.data.ProfileStorageData
 
 object PetAPI {
