+++ src/main/java/at/hannibal2/skyhanni/api/SlayerAPI.kt
@@ -1,5 +1,7 @@
-package at.hannibal2.skyhanni.data
+package at.hannibal2.skyhanni.api
 
+import at.hannibal2.skyhanni.data.ProfileStorageData
+import at.hannibal2.skyhanni.data.ScoreboardData
 import at.hannibal2.skyhanni.events.LorenzChatEvent
 import at.hannibal2.skyhanni.events.LorenzTickEvent
 import at.hannibal2.skyhanni.events.SlayerChangeEvent
