+++ src/main/java/at/hannibal2/skyhanni/events/SackChangeEvent.kt
@@ -1,6 +1,6 @@
 package at.hannibal2.skyhanni.events
 
-import at.hannibal2.skyhanni.data.SackAPI
+import at.hannibal2.skyhanni.api.SackAPI
 
 class SackChangeEvent(
     val sackChanges: List<SackAPI.SackChange>,
